add: go home with delay hook
