+++ src/components/pages/Signout/SignoutPage.tsx
@@ -1,24 +1,15 @@
 import type { NextPageWithLayout } from '@/pages/_app';
 import Link from 'next/link';
-import { useRouter } from 'next/router';
 import { useTranslation } from 'next-i18next';
-import { useEffect } from 'react';
 import BaseLayout from '@/components/layout/BaseLayout';
 import SignInOutBtn from '@/components/ui/Button/SignInOutBtn';
+import useGoHomeWithDelay from '@/hooks/useGoHomeWithDelay';
 import NetflixLogo from '@assets/netflix/top-logo.svg'
 import { SignoutBgCss, SignoutContentLayoutCss, SignoutDesc1Css, SignoutDesc2Css, SignoutGotoMainBtnCss, SignoutHeaderCss, SignoutLogoCss, SignoutSignInOutBtnCss, SignoutTitleCss } from './style';
 
 const SignoutPage: NextPageWithLayout = () => {
   const { t } = useTranslation(['common', 'page-signout'])
-  const router = useRouter()
-  useEffect(() => {
-    const goHomeTimeout = setTimeout(() => {
-      router.push('/')
-    }, 30000)
-    return () => {
-      clearTimeout(goHomeTimeout)
-    }
-  })
+  useGoHomeWithDelay(30000)
   return <div css={SignoutBgCss}>
     <nav css={SignoutHeaderCss}>
       <Link href="/" css={SignoutLogoCss}>
